refactor: split oversized production modules
Extract per-action verifiers from verification.rs into verification_actions.rs,
metric types from executor.rs into executor_metrics.rs, and RuntimePorts from
runtime/mod.rs into ports.rs. All six files are now ≤400 lines.
